Reorganizing models and config
