+++ src/shared/models/index.ts
@@ -1,0 +1,12 @@
+export type { RegisterPayload } from './authModels';
+export type { RegisterResult } from './authModels';
+
+export type { DocumentType } from './documentModels';
+export type { BarDataItem } from './documentModels';
+
+export type { LinkFormValues } from './linkModels';
+export type { LinkDetail } from './linkModels';
+export type { LinkData } from './linkModels';
+
+export type { User } from './userModels';
+export type { Contact } from './userModels';
